Do not truncate OS information in agent_control
When getting agent information from agent_control it provides really
useful OS information about the agent it is running on. Unfortunately,
that information gets truncated both when displaying in it's normal mode
and when CSV output is chosen.

At the very least the CSV output should show this data without getting
truncated. My guess is the data was originally truncated to make the
display nicer. That isn't a concern with CSV output.