More cleanup of code comments
Cleanup comments in an attempt to get some consistency. A good example
of comments I hate are those in read-agents.h where about each function
prototype the function and it's arguments were basically redeclared in
the comment. Seems kind of pointless to me since the same thing is a
single line below. If someone were to change that prototype and forget
to change the comment than it becomes even more pointless.