Update ssh_asa-fwsmconfig_diff
Changed 'send "show running-config"' to 'send "more system:running-config"'.  This allows OSSEC to detect changes to preshared keys, such as RADIUS and Tunnel-Groups.  Revealing this information allows OSSEC to do a more thorough job of inspecting the running configuration for changes.  Previously, this information was masked behind five asterisks.

Reversed the order of the output and inserted two sleep commands.  This eliminated false positive alerts due to an inconsistent number of line feeds at the end of the file.  Adding two sleep commands and moving the larger portion of the data to the top of the file has eliminated the issue in my testing.