Bugfixes for OS_IPFound, OS_IPFoundList, OS_IsValidIP.
OS_IsValidIP violated it's `const` constraint when the ip_addr element
was set to 'any'.  This caused some unexplained behavior when subsequent
calls.  This is corrected.

OS_IPFound and OS_IPFoundList where call sacmp(), but if OS_IsValidIP
was called with "any", the sockaddr_store.ss_family would be AF_INET6.
AFAICT, this meant any comparison of "any" to an IPv4 address would
fail.  This affected the ossec-remoted when validating IPv4 keys issued
by authd which set 'any' as the IP address to verify.

This would also cause failures *everywhere* in the rules, analysis,
exec, and activeresponse code where the src/dst IP being compared to any
was IPv4.