Make rootcheck sleep interval configurable
Sleeping for 2 seconds after each PID may be very slow on a busy system
(and during this time, syscheckd is not processing new events), and in
some cases it might be preferable to use other means of throttling (e.g.
cgroups) to limit rootcheck's impact on the system.

This flag lets the user customize rootcheck's sleep interval however
they want it (in our case, we're using 0 seconds and cgroups).