Corrected some minor compatibility issues between BSD and Linux in os_net.c for the ip multiplexing modification.
It turns out that getnameinfo() is much pickier on BSD than it is on Linux,
and it returns EAI_FAIL if the exact length returned by getaddrinfo() in the
sa->sa_len parameter is not passed in the call to getnameinfo().

Most versions of Linux do not have sa_len as part of the addrinfo struct so
you can just pass sizeof (sa), but if you do this on BSD it fails with
EAI_FAIL (error 4).

Also changed the counter being evaluated for socket creation. Under some
circumstances, socket 0 is a valid socket.  Instead of looking at the socket
vlue, we count the number of sockets and use that for comparicon instead.