A process to do dns lookups and connects.
Be able to fork a process to do dns lookups. This process will not
be chrooted, but will run as the ossec user.
It receives communicates with other daemons via imsg(3) from the
OpenBSD project. A dnslookup is performed, and a connect call is made.
If it succeeds, the fd is sent back to the calling process.
There is probably a lot of debugging stuff in this right now, and
it's very smtp specific. I hope to clean this up later. I also have
not tested this with ipv6 at all.