Don't use a variable when you mean another variable.
The logic changed a bit at some point, and the code didn't keep up
with the change. I was originally using a_sock for the socket, but
eventually just used agt->sock. Change the code to reflect that logic
change.
This is officially passing messages to the ossec server now.