analysisd: fix possible null ptr deref in OS_CleanMSG.
If the `msg` provided to `OS_CleanMSG` has a `(` after the ID, but
doesn't not contain a `->` or `:` it should be rejected with
a `FORMAT_ERROR`.

Prev. to this commit nesting `strstr` for `->` as the first argument to
`strchr` for `:` results in a null ptr deref when the message is
malformed.