analysisd: fix off-by-one in OS_CleanMSG.
When removing the '[ID xx facility.severity]' substring of a syslog
message in `OS_CleanMSG` care needs to be taken to stay within the
bounds of the msg buffer when advancing past the `] ` in the msg.

Prev. to this commit the `pieces` pointer is incremented by 2 when there
may be only 1 character remaining before the null terminator.