analysisd: fix heap overflow in rootkit decoder.
The `RK_File` function of the rootcheck decoder needs to ensure it
doesn't index outside of `MAX_AGENTS` when trying to find/open
a rootcheck file for a given agent.

Prev. to this change the value of `i` could exceed `MAX_AGENTS`
resulting in a heap buffer overflow accessing `rk_agent_ips[i]` or
`rk_agent_fps[i]`.

The fix is adapted from similar logic in the syscheck decoder, added as
a response to a matching vulnerability in that decoder patched in 2012.