os_regex: refuse to compile empty PCRE2 pattern.
The `OSPcre2_Compile` function should refuse to compile empty patterns
and instead return an error. This avoids an off-by-one error that can
occur later in the function when an assumption is made that the pattern
is always >= 1 characters long.

A distinct error code `OS_REGEX_PATTERN_EMPTY` (10) is added to
`os_regex/os_regex.h` to be returned by `OSPcre2_Compile` for empty
patterns similar to the `OS_REGEX_PATTERN_NULL` error.