etc/decoder.xml: remove empty program_name_pcre2 element.
The `os_regex` PCRE2 code has been updated to return an error when asked
to compile an empty pattern. As a result the empty
`<program_name_pcre2>` element in the example `decoder.xml` config for
the `"pam"` decoder needs to be removed or it will prompt an error from
`ossec-analysisd` when it tries to compile the PCRE2 pattern.