Being osdns() into the mail directory.
Trying to re-use the code in the way I did was a mistake. This
duplicates it, but significantly reduces the complexity.
While here, pass the mail config directly to the osdns() daemon.
Now this information doesn't have to pass in the imsg itself. Less
memory stuff going on.
Still needs a lot of cleanup.